chore: refactor chain-adapters type exports
