+++ packages/chain-adapters/src/bitcoin/index.ts
@@ -1,1 +1,2 @@
 export * from './BitcoinChainAdapter'
+export * from './types'
