+++ packages/chain-adapters/src/cosmossdk/cosmos/index.ts
@@ -1,1 +1,2 @@
 export * from './CosmosChainAdapter'
+export * from './types'
