+++ packages/chain-adapters/src/cosmossdk/osmosis/index.ts
@@ -1,1 +1,2 @@
 export * from './OsmosisChainAdapter'
+export * from './types'
