+++ packages/chain-adapters/src/ethereum/index.ts
@@ -1,1 +1,2 @@
 export * from './EthereumChainAdapter'
+export * from './types'
