+++ packages/chain-adapters/src/ethereum/types.ts
@@ -1,4 +1,4 @@
-import { AssetBalance } from '.'
+import { AssetBalance } from '../types'
 
 export type Account = {
   nonce: number
